third_party/WebKit/LayoutTests/http/tests/appcache/resource-redirect.html

<html manifest="resources/resource-redirect.manifest">
<script src="/js-test-resources/js-test.js"></script>
<script>
var jsTestIsAsync = true;
description("Test that a redirect makes resource caching fail.");

function unexpectedCallback(eventName)
{
    testFailed("An unexpected " + eventName + " event was dispatched on window.applicationCache.");
    finishJSTest();
}

function test(e)
{
    shouldBe("window.applicationCache.status", "window.applicationCache.UNCACHED");

    event = e;
    shouldBeEqualToString("event.reason", "resource");
    shouldBeEqualToString("event.url", "http://127.0.0.1:8000/appcache/resources/resource-redirect.php");
    shouldBe("event.status", "307");
    shouldBeTrue("'message' in event");

    finishJSTest();
}

applicationCache.addEventListener('cached', function() { unexpectedCallback("cached"); }, false);
applicationCache.addEventListener('noupdate', function() { unexpectedCallback("noupdate"); }, false);
applicationCache.addEventListener('error', test, false);

</script>
</body>
</html>
