third_party/WebKit/LayoutTests/http/tests/appcache/resource-redirect-expected.txt

Test that a redirect makes resource caching fail.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.applicationCache.status is window.applicationCache.UNCACHED
PASS event.reason is "resource"
PASS event.url is "http://127.0.0.1:8000/appcache/resources/resource-redirect.php"
PASS event.status is 307
PASS 'message' in event is true
PASS successfullyParsed is true

TEST COMPLETE
